ciri: move conf to ciri folder
