--- ciri/conf/BUILD
@@ -1,21 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-cc_library(
-    name = "usage",
-    srcs = ["usage.c"],
-    hdrs = ["usage.h"],
-)
-
-cc_library(
-    name = "conf",
-    srcs = ["conf.c"],
-    hdrs = ["conf.h"],
-    deps = [
-        ":usage",
-        "//ciri/api:conf",
-        "//common:errors",
-        "//consensus:conf",
-        "//gossip:conf",
-        "//utils:logger_helper",
-    ],
-)
